Removed the get_price responsibility from DataProvider and moved it to AbstractPriceDataProvider
