+++ qf_lib/data_providers/preset_data_provider.py
@@ -31,11 +31,11 @@ from qf_lib.containers.futures.future_tickers.future_ticker import FutureTicker
 from qf_lib.containers.qf_data_array import QFDataArray
 from qf_lib.containers.series.prices_series import PricesSeries
 from qf_lib.containers.series.qf_series import QFSeries
-from qf_lib.data_providers.data_provider import DataProvider
+from qf_lib.data_providers.abstract_price_data_provider import AbstractPriceDataProvider
 from qf_lib.data_providers.helpers import normalize_data_array
 
 
-class PresetDataProvider(DataProvider):
+class PresetDataProvider(AbstractPriceDataProvider):
     """
     Wrapper on QFDataArray which makes it a DataProvider.
 
@@ -357,3 +357,13 @@ class PresetDataProvider(DataProvider):
     def _adjust_end_date(end_date: Optional[datetime]) -> datetime:
         end_date = end_date or datetime.now()
         return end_date + RelativeDelta(second=0, microsecond=0)
+
+    def price_field_to_str_map(self) -> Dict[PriceField, str]:
+        pass
+
+    def expiration_date_field_str_map(self, ticker: Ticker = None) -> Dict[ExpirationDateField, str]:
+        pass
+
+    def _get_futures_chain_dict(self, tickers: Union[FutureTicker, Sequence[FutureTicker]],
+                                expiration_date_fields: Union[str, Sequence[str]]) -> Dict[FutureTicker, QFDataFrame]:
+        pass
